requirements.txt
streamlit
streamlit_option_menu
firebase-admin
pyrebase4